feat: implement delete functionality on frontend and connect to backend controller

+++ server/controller/drink.controller.js
@@ -16,7 +16,7 @@ drinkController.createDrink=async(req,res)=>{
         } = req.body;
         const newDrink = new Drink({category, name, size, caffeineContent}) //import model
         await newDrink.save()
-        res.status(200).json({status :"ok", data:newDrink}) //instead of send send status , The syntax here is using method chaining, which allows you to call multiple methods on the same object in a single line
+        res.status(201).json({status :"ok", data:newDrink}) //instead of send send status , The syntax here is using method chaining, which allows you to call multiple methods on the same object in a single line
     }
     catch(err){
         res.status(400).json({status:"fail", error:err})
@@ -37,6 +37,10 @@ drinkController.getDrink=async(req,res)=>{
                 $lt: endOfDay     
             }
         }).select("-__v"); // exclude __v field
+
+        if (drinkList.length === 0) {
+            return res.status(404).json({ status: "fail", message: "No drinks found for today" });
+        }
         res.status(200).json({status:"ok", data:drinkList})
     }
     catch(err){
@@ -53,6 +57,11 @@ drinkController.updateDrink=async(req,res)=>{
         const fields = Object.keys(req.body);
         fields.map((item)=> (drink[item]=req.body[item]));
         await drink.save;
+
+        if (!drink) {
+            return res.status(404).json({status: "fail", message: "Drink not found"});
+        }
+
         res.status(200).json({status:"ok", data:drink});
     }
     catch(err){
@@ -63,6 +72,11 @@ drinkController.updateDrink=async(req,res)=>{
 drinkController.deleteDrink=async(req,res)=>{
     try{
         const deleteItem = await Drink.findByIdAndDelete(req.params.id);
+
+        if (!deleteItem) {
+            return res.status(404).json({status: "fail", message: "Drink not found"});
+        }
+
         res.status(200).json({status:"ok", data:deleteItem})
     }
     catch(err){
